Remove path-checking in images' resource configs
Having a resource that restricts the path that gives us versions for
`golang-builder-dockerfile`, we end up without the necessary files to
properly running the task.

Adding the task configuration to the list of path helps, but then we
miss the configuration for the notification step, meaning that we'd then
need to have another `get` just for that.

As there seems not to be problematic to constantly build
`concourse/golang-builder` on every push to `ci` anyway, we removed the
path-specific resources in order to ensure that we don't end up using
old configurations.

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>
Co-authored-by: Joshua Winters <jwinters@pivotal.io>