add restore to backup pipeline
As very well pointed out by @pivotal-jamie-klassen, backups are only as
good as their restore.

This commit represents an effort to have that second half (the restore
part) as the subsequent job right after the backup.

In overall, the pipeline for backup + restore looks like this:

	BACKUP

		1. take a backup from prod using BBR
		2. store the backup in GCS

	RESTORE

		1. generate ops files that match the versions in prod
		2. deploy concourse having just a database
			2.1 with `superuser` to concourse so we can enable
			    `pg_statements`
		3. `bbr restore` against it
		4. pause all pipelines
		5. bring `web` and `worker` up
		6. trigger a job in a pipeline that already existed before

This way, we're able to verify not only that `bbr` is able to put the
bits back into a database, but that concourse is able to use it
accordingly too.

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>
Co-authored-by: Bishoy Youssef <byoussef@pivotal.io>
Co-authored-by: Jamie Klassen <cklassen@pivotal.io>
Co-authored-by: Zoe Tian <ztian@pivotal.io>
Co-authored-by: Bin Ju <bju@pivotal.io>
Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>