wip: add compact `concourse` pipeline for ubuntu-base concourse image
With the goal of having a `concourse/concourse` image that is based of
Ubuntu-based resources types, we end up refactoring the pipeline in a way
that would allow us to consume them without getting into hacks like we
did for the BOSH release where we'd swap contents from the tarball with
resource types that we'd get from GitHub.

This commit contains our first iteration on a minimal pipeline that
contains just those jobs and resources that interact with the image and
its resource types.

The approach that we went for is:

	1. separate `dev-image` from `resource-types`

	2. consume in `dev-image` the `alpine`-based resource types

	3. produce `alpine` **and** `ubuntu` linux-rc's
	  - this way we keep track of the exact tarball with the resource types
	    that got to the BOSH release
	4. produce `alpine` **and** `ubuntu` `concourse/concourse` images

	  - this way those not wanting to leverage `ubuntu` don't get hurt by
	    the size of `ubuntu`-based resource types

	5. generate a list of the debian packages that are present in the final
	   image
	  - so we can scan for non-compliant packages and CVEs

This will get rid that ugly bunch of resource types getting dumped into
`bosh-bump` (where we do the resource types swapping).

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>
Co-authored-by: Bishoy Youssef <byoussef@pivotal.io>
Co-authored-by: Bin Ju <bju@pivotal.io>
Co-authored-by: Zoe Tian <ztian@pivotal.io>
Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>