add separate flavors of linux-rc and image to concourse pipeline
As a follow up to the previous commit that added the `concourse-compact`
pipeline that allowed us to iterate quickly on the jobs that touches the
creation of different images and linux-rcs, this merges that pipeline
with `concourse.yml` (the main one).

In order to not duplicate the amount of testing we have for anything
image-related, we take the approach of going `-alpine` based on
everything that is test related, taking care of ubuntu only at the
publishing step.

If we decide to switch, it's just a matter of replacing
`input_mapping`s.

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>