concourse: remove passed constraints for `unit-image`
Previously, in order to have a job in the pipeline to leverage an
updated `unit-image`, you'd need to have the whole pipeline successfully
run in order to benefit from that update.

That made the iteration slow.

By adopting the same pattern as `ci`, we're able to independently update
`unit-image`, just like `ci`.

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>
Co-authored-by: Bin Ju <bju@pivotal.io>