release: pin charts version to 5.2
As the Chart contains the fields that are specified for any given
version (e.g., `master` currently points to 5.4.1), when leveraging
`branch: master` in the release pipeline for 5.2., we end up with
problems given that the fields differ.

Thus, this commit pins the version of `helm/charts` to the last commit
where the concourse versions match with the Chart.

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>
Co-authored-by: Bin Ju <bju@pivotal.io>