concourse: fix installation of `gdn` to not include version
This is intended to fix the problem that
https://github.com/concourse/ci/pull/98 introduced when putting `gdn` to
be fetched directly from the `gdn` github-release in the `build-rc`
step: rather than coming named as `gdn`, it now comes as `gdn-$VERSION`.

As in later steps we trust that guardian is well-named (i.e., named
`gdn`) and is executable, we used to make any jobs after `build-rc`
fail.

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>
Co-authored-by: Bin Ju <bju@pivotal.io>