temporarily replace `apache` go mod dependency
As of literally right now (Sat 31 Aug 2019 20:02:43 EDT), git.apache.org
is facing troubles, and is not expected to get up very soon.

To get around that, this PR sneaks in a replacement directive that
modifies the source of the thrift code when we're fetching dependencies
to run our tests.

Note.: this is meant to be TEMPORARILY (like, really temporarily - an
issue to revert it when they get back up again was created).

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>