dockerfiles/dev: add `containerd`
coming along with https://github.com/concourse/concourse/pull/4752, this
commit adds `containerd` to our base (`concourse/dev`) so that we can
leverage `containerd` in `concourse/concourse:local` and other places in
the pipeline.

note that differently from `gdn`, we're not taking `containerd`, `runc`
and `cni` artifacts from the filesystem (populated by a pipeline) -
that's because there's quite a bit of version interdependency between
those that having them.

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>