Fix bosh-smoke for release pipelines
bosh-smoke no longer works for release pipelines because this job
assumes that it is running on a vm that is inside a bosh environment.
Failure occurs at the last step of the job, `smoke`, where we try and
reach the atc using an internal bosh domain. This fails because the job
now runs from a k8s cluster.

The changes in this commit run bosh-smoke on the `topgun` worker, which
is running inside a bosh bbl environment.

Signed-off-by: Taylor Silva <tsilva@pivotal.io>