Update helm-pr pipeline for new concourse-chart repo
There are now multiple helm-pr pipelines:

- `helm-prs` is for testing commits against master. It tests the latest
stable concourse/concourse image with the k8s-topgun it was released
with
- `helm-prs-x.x.x` pipelines for when we need to backport patches.
These allow for fast feedback.

We also tried running a smaller subset of the k8s-topgun suite that
covered a majority of user-facing features. We chose the "Worker
Lifecycle" tests which are 2/43 of all tests. These tests alone take
9-10mins to run vs 18-20mins for the entire suite. For now we'll run the
entire test suite. In the future we could create a "smoke" test suite
that runs faster (2-3mins?) potentially and use that for testing PR's
intead.

Signed-off-by: Taylor Silva <tsilva@pivotal.io>
Co-authored-by: Bin Ju <bju@pivotal.io>