pipelines/concourse: add `build-integration-image` to groups
given that we make use of groups[1] for all other jobs that we have, we
must include `build-integration-image` too, otherwise the pipeline can't
be set.

here I add it to the `images` group as its main purpose is to build an
image to be used in the rest of the pipeline (just like we do for
others).

[1]: https://concourse-ci.org/pipelines.html#groups

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>