pipelines/concourse: add containerd-integration
As the `containerd` backend integration tests require `containerd` to be
running, we need to have its binaries present in the image that forms
the base that runs such tests.

Given that `unit` has been the place where all extra dependencies get
added, there's where `containerd` end up being added to.

For the execution of the tests, a separate job was created so that we
don't get possible failures from `testflight` when the problem is
isolated to `containerd` only.

fixes concourse/concourse#4878

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>