dockerfiles/unit: pass custom config to containerd
as we're running `containerd` within a container (whose rootfs is
overlay-based), and `containerd` is supposed to leverage overlay for its
snapshotting facilities, we need to direct containerd to use a
non-overlay based directory (`/scratch`).

given that we're not running the containerd process as the runtime for
kubernetes (through the CRI[1]), we can disable the CRI plugin.

see [2] for the config format.

[1]: https://github.com/containerd/cri
[2]: https://github.com/containerd/containerd/blob/80d50a8a636d0961571b93679956615b66c69e4d/docs/man/containerd-config.toml.5.md#format

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>