dockerfiles/containerd: force use of concourse's containerd binaries
Previously, `unit` had PATH configured with higher priorities in the
search for containerd binaries to those paths that were not
`/usr/local/concourse/bin`. This was to that when `docker` comes up (or
would create a container), it'd pick the right binaries for its version.

By doing so, when running our own version of `containerd`, we'd end up
picking the binaries that suited `docker`'s version, not ours.

Thus, having our `containerd` starting with PATH set to the place where
we know our binaries exist, forces it to find the right binaries.

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>