use set-pipeline step in reconfigure pipeline
Also within the prs pipeline, when we added skip download to the get
after the puts for updating the pr status, that actually made it so that
the volume passed to the task after was empty. This happens because
skipping download still created a get step but since we skipped
downloading, the get volume is empty and that overwrites the original
volume. In order to avoid this, I added an alias to the put step so that
the empty volume will not overwrite the concourse-pr volume.

Signed-off-by: Clara Fu <cfu@pivotal.io>