add version bump checking, linting, and install test
We had a PR go in that somehow got through k8s-smoke and k8s-topgun,
both which set a lot of variables in the chart with --set, and broke the
default install of the chart (doing `helm install concourse/concourse`
failed).

This install test mimics the install shown in the chart README.md and
ensures coverage of that one use-case which first time users of the
chart all use. It's the chart's version of `concourse quickstart`. We
were not testing this automatically anywhere.

Signed-off-by: Taylor Silva <tsilva@pivotal.io>