never use external linker to build fly
We originally started using an external linker to build fly when we
were depending on go 1.12. It seems that this is no longer necessary
now that we are using go 1.14.

The key feedback mechanism that I can recall us relying on that guided
us to introduce this parameter was the following: create a GCE instance
running CentOS 6 (which runs Linux 2.6), and execute fly on it. For
thoroughness, we repeated this manual test successfully with the changes
in this patch.

Signed-off-by: Jamie Klassen <cklassen@pivotal.io>
Co-authored-by: James Thomson <jthomson@pivotal.io>
Co-authored-by: Ciro S. Costa <cscosta@pivotal.io>