download fly cli in testflight for release pipelines
Rather than building `fly` in testflight for our release pipelines, we
should be downloading the CLI from ATC. This lets us exercise the actual
fly binary that gets bundled with Concourse (and can help us detect bugs
with the flags we user to build fly early)

For our PR pipeline, we do not want to download fly, and instead want to
build it. This is because a PR may make changes to fly, and we want to
run the tests against the updated code (rather than what was baked into
the `concourse/dev` image).

Extends concourse/concourse#5266

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>
Co-authored-by: Rui Yang <ryang@pivotal.io>