topgun: add uaa (release and concourse pipelines)
in https://github.com/concourse/concourse/pull/5255, `uaa` was added to
the set of releases necessary for a complete run of `topgun`.

these changes allows us to keep track of the releases of UAA that we're
using when running those tests, as well as allowing our task to upload
that to the BOSH director.

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>