release: slack-notify on build errors
Recently some inputs were added to the `build-dev-image` task, in the
interest of continuously updating `containerd`'s dependencies. Unfortunately,
the job where this task runs was only updated in the main pipeline and the
release pipeline was left out-of-date.

Accordingly, builds of the `dev-image` job in the release pipelines were
erroring, but this did not cause the exising `on_failure` hook to trigger, so
we were not being notified of these problems. This could potentially allow
the release pipelines to continue to drift out of date and not be ready for a
quick release, say, in the event of a security incident.

This change adds a notification even in cases when configuration problems
prevent the build from fully executing.

Signed-off-by: Jamie Klassen <cklassen@pivotal.io>