Add USE_CONTAINERD variable to smoke deployment
When deploying a smoke env using terraform we want the ability to
enabled containerd on the worker using systemd.

Any environmet variables that start with TF_VAR_ are interpolated by
terraform and can be used in your tf files. These vars are
case-sensitive.

The workspace still needs to be different from the existing bin-smoke
job to ensure that both jobs are using different vm's.

Signed-off-by: Taylor Silva <tsilva@pivotal.io>
Co-authored-by: Rui Yang <ryang@pivotal.io>