Update how helm-prs checks for the version bump
The git and egrep commands were not fool proof with always ensuring that
the chart version was bumped upwards. If someone moved the version
number backwards then it would have passed this check.

This golang script ensures that the version number is always increased
and doesn't include any extra metadata o prerelease information

Signed-off-by: Taylor Silva <tsilva@pivotal.io>