do not set env vars for keys before they exist
We are in the process of switching release/5.5.x to depend on vito's fork of
go-flags. On this fork, flags passed to the `concourse` binary via env vars are
validated. Part of the Dockerfile in concourse/concourse has steps where
`concourse generate-key` is run, and these steps will fail if they are run with
key-related flags set to nonexistent paths via env vars as they were here.

Signed-off-by: Jamie Klassen <cklassen@pivotal.io>