use full semver in RC binary filenames
concourse/ci#307

The new format for an RC binary's filename is

concourse-<major>.<minor>.<patch>-rc.<rc>+<commit>.<platform>.amd64.<ext>

where platform is one of:
* linux (where the bundled resource types are based on alpine)
* linux-ubuntu (where the bundled resource types are based on ubuntu)
* darwin
* windows

and ext is `.tgz` except on windows where it is `.zip`.

What's great about this is that it is semver-compliant, so our gcs resources
will properly order their versions, and it gives us an easy way to track which
build produced a given binary.

Signed-off-by: Jamie Klassen <cklassen@pivotal.io>