Auto bump the chart version when publishing
This commit amends and duplicates one job. The only difference between
the jobs is whether they publish a new MAJOR or MINOR version of the
concourse helm chart.

A commit is made in the package-chart task and that commit is eventually
pushed by the CI bot in the final put step.

Most of this change was taken from this commit:
22456bc6d80ae7096880f375417e640bf263092b

Signed-off-by: Taylor Silva <tsilva@pivotal.io>