Add NORSK_PIPELINE to task that finds status.json for a given pipeline
The script that was reading status.json from oslo-scan-status repo,
was always reading it from the p-concourse folder, instead of reading
the folder that corresponds to the pipeline name (e.g. p-concourse-5.5.x).

Now the task takes in the pipeline name and uses it in the path to status.json
file.

Signed-off-by: Izabela Gomes <igomes@pivotal.io>