Using different formatting slack message script for norsk related tasks
The current format-slack-message script tries to find a person responsible for
fixing/breaking changes using `yarn blame`, which is not suited for our
norsk scenario. The job looks at the oslo-scan-status repo to find out about
failed jobs, but that repo gets updated by a bot (so, no specific person to
blame).

The new script will instead look at status.json for failed jobs and save the
urls (of failed jobs from the norsk pipelines) to be sent to a slack channel.

Signed-off-by: Izabela Gomes <igomes@pivotal.io>