separate reconfigure.yml to team main and contibutor
For security contributors should be limited to operator PRs
pipeline only. With current implementation of Concourse RBAC
configuration it can't do that if PRs pipeline stays in the
same team with others.

So this commit moves PRs pipeline from main team to contributor
team and adds a reconfigure pipeline for it.

Signed-off-by: Rui Yang <ryang@pivotal.io>