add containerd integration test to prs
This test suite was running in the concourse pipeline but missing from
prs. Skipping ./worker/runtime/integration in unit job as directory has
been renamed. There will be prs currently in the works that don't have
the renamed directory. They would then fail on unit job. Contributors
will likely be stumped and won't know to rebase. Therefore, keeping the original
./worker/backend/integration in the list as well to avoid unnecessary
failures. This should be removed in a few weeks.

Signed-off-by: James Thomson <jthomson@pivotal.io>
Co-authored-by: Muntasir Chowdhury <mchowdhury@pivotal.io>