Fix containerd integration tests
In concourse/concourse worker/backend is being renamed to worker/runtime.
There will be prs currently in the works that don't have
the renamed directory. They would then fail on this job. Contributors
will likely be stumped and won't know to rebase. Therefore, keeping the original
./worker/backend/integration in the list as well to avoid unnecessary
failures. This should be removed in a few weeks

Signed-off-by: Muntasir Chowdhury <mchowdhury@pivotal.io>
Co-authored-by: James Thomson <jthomson@pivotal.io>