retry watsjs tests
Despite my efforts in concourse/concourse#5627, watsjs is still somewhat
flaky. It typically passes after 1 or 2 retries though. So, why not just
retry it automatically in the pipeline?

I also got rid of the retries on `yarn-build` - if it doesn't compile
the first time, it's not going to compile when we retry it.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>