Map concourse-image to latest-concourse-image
We were seeing this error in the upgrade and downgrade tasks

```
ERROR: no such image: concourse/concourse:: invalid reference format
```

Seemed like the image tag for concourse/concourse was empty. The script
expects the image to appear in `concourse-image` but the resource was
renamed to `latest-concourse-image`. This mapping fixes the task.

Signed-off-by: Taylor Silva <tsilva@pivotal.io>