generating release notes through release-me
Added a create-draft-release job that will create a draft release with a
generated release note. The release note is generated using the
release-me CLI which grabs all prs merged after the last release of
Concourse. It will run automatically after a commit has made it through
the testing part of the pipeline. The draft release will then be
converted into a real release in the publish binary job.

[#5768]

Co-authored-by: Taylor Silva <tsilva@pivotal.io>
Signed-off-by: Clara Fu <cfu@pivotal.io>