edit create draft release job
Since the draft release job runs right before ship-it, it will generate
a release note only after a commit has made it through the pipeline. I
was worried that if a release note was generated right after a pr was
merged but hasn't made it through the pipeline, the release note would
still include that merged pr because the release note generator fetches
all prs merged after the latest release. So I editted the release me CLI
to take in a last-commit-SHA which will allow for it to generate a
release note with all prs merged before that last commit SHA. We can
pass in the commit that has made it through the testing part of the
pipeline into the last-commit-SHA so that the release note generator
will only generate a release note with commits that have made it past
the pipeline.

[#5768]

Signed-off-by: Clara Fu <cfu@pivotal.io>