differentiate between major/minor/patch bumps for chart
There are three main cases we want to cover. This commit covers what we
think are the two most common cases.

1. a new major/minor of Concourse has been released. In this case we'll
want to merge the dev branch into master as it contains the work related
to the new major/minor of Concourse
2. a new patch of Concourse has been released for the latest minor. In
this case we want to merg the release/x.x.x branch back into master.
There is likely already commits on dev related to the upcoming
major/minor release which should not be merged into master at this point

The third case is in regards to LTS versions of Concourse. For now the
bumping can be done manually on those release branches (there may be
more than one LTS!). We don't want to merge LTS changes back into master
unless the LTS also happens to be the latest OSS minor available,
therefore also falling into case #2 above.

Signed-off-by: Bohan Chen <bochen@pivotal.io>
Co-authored-by: Taylor Silva <tsilva@pivotal.io>