simplify chart workflows
We decided that this pipeline (helm-chart) should only concern itself
with the dev and master branches. It should not care about the
release/x.x.x branches. Those can be the conern of the release
pipelines. We honestly don't even care about tagging any releases coming
from the release branches in the concourse-chart repo. All releases come
from master already and the release branches are really there for LTS
releases. If a new version (major/minor/patch) of Concourse comes out
they all follow the same workflow on the master branch of
concourse-chart:

1. Bump the app version
2. if major/minor app version bump, then merge dev into master as well
(optional)

We basically realized that some Concourse releases require no changes to
the chart, therefore there's nothing on dev to merge into master. Other
times there are changes tied to a verison of Concourse that affect the
chart and THEN we want to merge dev into master.

Again, this pipeline should not are or concern itself with what's
happening on any of the release branches.

Signed-off-by: Bohan Chen <bochen@pivotal.io>
Co-authored-by: Taylor Silva <tsilva@pivotal.io>