make creating draft release a manual job
Because of the current releasing process being that the release anchor
will need to manually edit the draft release before shipping but if
there is a new commit that comes in, it will create a new draft release
and overwrite any manual edits that were made previously. So in order to
ensure that the anchor will only edit after they know this draft release
is final, we can have the creating of the draft release be a manual job.

Signed-off-by: Clara Fu <cfu@pivotal.io>