update terraform runtime selection logic
CONCOURSE_USE_CONTAINERD has been replaced by CONCOURSE_RUNTIME.
This unfortunately is going to break the bin-smoke-containerd
test for 6.x pipelines from 6.4 and below. containerd is
experimental in those releases and the job will be paused
until we figure out a nice way to support the old 6.x pipelines
and the current ones.

Signed-off-by: Muntasir Chowdhury <mchowdhury@pivotal.io>