set flags based on concourse version
This solves the problem of different versions of concourse requiring
different flags.

The current example is that < 6.5.0 the container
runtime flags was separated into `use_containerd=true`, `use_garden=true`, but >=
6.5.0, they were consolidated into `runtime=containerd`.

However, since the ci repo doesn't maintain versions, we need a way to figure out the
version for each pipeline and set different flags. This will be done in the
terraform-smoke script, which now has knowledge of what each version of
concourse needs

Signed-off-by: Bohan Chen <bochen@pivotal.io>