ensure resource types were tested in dev-image
Make sure the same version of resource-types are packaged for both
ubuntu and alpine, and make sure that version is the one that was
bundled in the dev-image. Although we only use the alpine resource types
in our dev image, add resource-types-ubuntu-image as a build input to
`dev-image` so that we can get `passed` on `dev-image`.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>