support building dev image in upgrade/downgrade
In order to use these tasks in the PR pipeline, allow building the dev
image in the upgrade/downgrade tasks.

Initially I tried using the oci-build-task to build the image explicitly
and pass it in to the tasks, but `oci-build-task` produces a `digest`,
not a `tag`, so more modification would be required in the task.

I also thought about having a single job that builds the image with
passed constraints to all the PR test jobs, but decided against it,
since we'd need to upload every image to docker hub.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>