stop worker before recreating deployment
This is necessary due to the way docker-compose recreates containers.
Without this, the ATC may continually try to use the old worker to run
the resource check, despite that old worker no longer existing. This is
because the old worker couldn't unregister itself in time as the ATC was
terminated before the worker.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>