properly cleanup namespaces
the default values on concourse chart creates 2 namespaces, $releasename
is the namespace for all the components like web, worker, db.

another namespace is also created for every single team that is
configured in the values file and this namespace is persisted even after
the helm release is uninstalled. By default the main team will generate a
$releasename-main namespace to store secrets.

this commit disables the persistence behaviour so that the team
namespace is cleaned up when the release is uninstalled. Finally, since
helm doesn't cleanup the component namespace, we have to delete that
manually

Signed-off-by: Bohan Chen <bochen@pivotal.io>