Change int to string to fix k8s-topgun on 6.3.x
We did a fly get-pipeline on release-6.3.x and found that the
`helm_version` variable, which we passed as an int, ended up getting
turned into an int when the pipeline was set. Changed everything to be a
string since we don't feel like fighting that battle right now about why
the int became a string.

Signed-off-by: Taylor Silva <tsilva@pivotal.io>
Co-authored-by: Clara Fu <cfu@pivotal.io>