archive 6.5.x pipeline and use helm2 for 6.6.x pipeline
6.6.x release pipeline does not include the changes to use helm3 so I
will be switching it back to use helm2 instead. All release pipelines
created from here on out should be able to use helm3.\

Also removed the 6.5.x pipeline from reconfigure so that it will be
archived. It is currently paused and I don't think we are meaning to
support it.

Signed-off-by: Clara Fu <cfu@pivotal.io>