stop building vms with legacy kernel
This turned out to be a lot harder than expected to get working. I was
never able to get kernel 4.15 working on ubuntu 16 (which is already
running 4.15....). I highly suspect that GCP requires some specific patches
applied to the kernel and that the stock kernel wouldn't work. See
https://packages.ubuntu.com/xenial/linux-image-gcp

In addition to that, other problems include
- it required a lot of manual work to copy the kernel files from
kernel.ubunut.com
- it was hard to specify a build matrix for kernel + os combinations
(e.g. kernel 4.4 and 4.15 with ubuntu 16, but 4.15 and 5.4 with ubuntu
18)

Instead, let's just not worry about legacy kernels, and instead do smoke
testing on lts versions.

Signed-off-by: Bohan Chen <bochen@pivotal.io>