explicitly add git/mingw to $env:Path
honestly, I can't tell why Chocolatey's shims or $env:Path updating
isn't working; it works when I RDP in to make a new shell, but doesn't
work when I 'refreshenv' in a 'fly intercept'.

note: previous version of this also added Go to the $env:Path, but that
part seems to be working fine, so let's just add git and mingw.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>