set prs pipeline from the release pipeline
this could have been done in reconfigure pipeline, but then we need to
make sure the release pipelines we set stay in sync with the prs
pipelines we set, which is annoying.

rather than interpolating the branch into the `set_pipeline` name (which
is verboten), use `instance_vars` to construct an instance group of prs
pipelines - this is the reason set_pipeline name interpolation was
disallowed in the first place! plus, I'm curious to see instance groups
in action with real pipelines

p.s. when we have prototype var_sources and dynamic across step, we'll
set an instanced pipeline for *each* PR, rather than an instanced
pipeline for each branch. it'll probably have two instance vars at that
point - base branch and pr number

another important note - we'll have to enable instanced pipelines on CI
to get this to work

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>