go back to using image: for fly-build-linux task
just fetch it in rootfs format too. if and when we standardize on OCI
format (fingers crossed w/ containerd) the extra 'get' can be removed.

also remove USE_EXTERNAL_LINKER param as it's no longer a thing, and
remove some params which are already set in the task config file that
the step runs

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>