add job to upload releases for bosh tests
The bosh smoke and topgun tests are required to upload the same releases
to bosh and sometimes they can fail if they try to upload the same
releases at the same time (failing to acquire the lock for the release).
This commit adds a job that will upload the releases ahead of time so
that when each of the bosh tests run, they will no longer need to upload
them themselves.

Co-authored-by: Alex Suraci <suraci.alex@gmail.com>
Signed-off-by: Clara Fu <fclara@vmware.com>