fix reconfigure-teams
it was still looking for an input named ci.

we also realized we need two inputs - one for the task script, and one
for the team config. this is because the team config resource won't get
commits that just update the task

this is imperfect because if a commit modifies both the team config and
tasks, the pipelines-and-tasks resource may not have found that commit
yet, and will run with the old task. but, not sure there's a good
workaround

Signed-off-by: Bohan Chen <bochen@pivotal.io>
Co-authored-by: Aidan Oldershaw <aoldershaw@pivotal.io>