Fix quickstart-smoke
The override file was using our old concourse/concourse-rc image, which
we haven't pushed to in forever. It's now concourse/concourse-dev.

Docker hub finally deleted the image which is why this job is failing
now. Moved the override file to this repo, where we actually use it,
instead of keeping it in the docs repo.

Signed-off-by: Taylor Silva <tsilva@pivotal.io>