use timestamp for dev version
in the main (concourse) pipeline, rather than using the number of
commits since the last tag. this way, we're always rolling forward, and
build-image will be triggered on changes to things besides new commits
to the concourse repo (e.g. a new dev-image) - previously, the only
factor of "newness" was the commit to concourse/concourse.

this also moves the logic to construct the dev version into its own
task, allowing us to simplify the build scripts

Signed-off-by: Esteban Foronda <eforonda@vmware.com>
Co-authored-by: Aidan Oldershaw <aoldershaw@pivotal.io>