add branch pipeline
this pipeline is for testing/building a branch in the concourse repo
with the intention of testing large changes on a real environment prior
to merging. it doesn't run any of the k8s/bosh tests. it has a (manually
triggered) bump-prod job which updates concourse/infrastructure with a
passing version from the branch, so we can update CI easily.

also planning on adding auto-rebasing against master

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>