add task for updating the resource type versions
this will be run after shipit. rather than reading the
reconfigure-pipeline in a more structured way (e.g. parsing the entire
YAML document), I decided to use text-based traversal and only parse the
individual step. the reason for this approach was that I didn't want to
clobber any comments in the pipeline.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>