release: passed constraints include build-rc-image
this way, the rc and the rc image must refer to the same bundle.

we noticed that there was a bug here when we inadvertently shipped 7.3.0
with a broken version of gdn - it didn't pass k8s-smoke, but we were
still able to ship since it did pass the bosh jobs.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>