provide service account key for k8s-topgun auth
previously, the kube_config we provided to tasks that tested against the
k8s topgun cluster used basic auth to authenticate with the cluster. GKE
deprecated this, however (see concourse/infrastructure#12).

instead, the approach we're taking is to provide a kube_config to each
task that has no credentials, but instructs kubernetes to authenticate
using a GCP service account. we also provide the service account key and
write it to disk, and set the GOOGLE_APPLICATION_CREDENTIALS env var so
kubectl can locate it.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>
Co-authored-by: Esteban Foronda <eforonda@vmware.com>