set deploy PRs
track the list of PRs with the label "deploy" separately. compute the
list of PRs that are not labeled with "deploy", and set them as normal
(i.e. with the regular pr.yml pipeline). the ones labeled as "deploy"
are set with the special pr-deploy.yml pipeline

note that we're only doing this for the master branch (i.e. not the
release branches) - it's possible to do, but requires changes to
release.yml as well (and i'm not sure how useful it'd be).

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>