cross-compile fly and concourse in PR pipeline
rather than using the windows/darwin machines to compile fly and
concourse on PRs, now that we no longer rely on CGO (for the most part -
see (1)), we can cross-compile the binaries on linux. because of the
caveat mentioned in (1), I added a special task for cross-compiling fly
rather than modifying the existing one (which sets CGO_ENABLED for
darwin)

if this works well, we can cross-compile concourse on linux, which
should reduce the load on our darwin/windows workers, which tend to be
less stable when a bunch of workloads come in (there's only one of each)

(1) we actually *do* still rely on CGO for the darwin build. it'll
compile without it, but the Go DNS resolver seems to break split DNS for
the

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>