Use semver_constraint to track resources in release pipeline
This is how we used to track the github releases and figured we could do
the same here with registry-image.

I wasn't sure if it would detect sha changes to the same tag so I did
some testing to verify that it would pick up the daily images that the
resoures ship. It does so we're good to go back to the "old" behaviour

Signed-off-by: Taylor Silva <tsilva@pivotal.io>