fix alpine/ubuntu package pollution
Noticed this bug in production CI that linux-dev-alpine and
linux-dev-ubuntu has versions disapear after build-concourse job.
Turn out it's the file generated by building dev-ubuntu overwrite the
one for alpine since the "platform" we passing in for both tasks are the same
"linux". So the final file name doesn't match the regex of either
above resources.The versions will short live after the put steps
of build-concourse, then it will disapear by GC. But the damage has been
done in our gcp buckets that alpine package is replaced by ubuntu one
(guess the ubuntu one runs slower and always be pushed after alpin one)

When this happens in release pipeline, the much bigger problem is
the linux tarball we have been releasing since this change
https://github.com/concourse/ci/commit/3407736ea73d9df7ac557b9b50938f19b37e6a8a
is always the ubuntu version. You can tell from the size of the
linux tarball of concourse v7.4.0 and v7.5.0, where the size changed
from 790MB to 1.2GB.

And...this has happened before :(
https://github.com/concourse/concourse/issues/4586

Signed-off-by: Rui Yang <ruiya@vmware.com>