no significant changes to mounts
the mode for a lot of files have inode64 on them now, not something we
control directly though.

cgroup/misc was added, we'd get that for free too without making any
changes.

looks like /proc/sched_debug isn't there in newer kernels. Struggled to
find docs about this, just seems to be gone.

/sys/devices/virtual/powercap is a new one, but doesn't seem to matter
if we mount that or not

Signed-off-by: Taylor Silva <dev@taydev.net>