add compile-check job to PRs pipeline
This job compiles the main concourse binary for all platforms. This is
to catch any obvious compile errors that may have arisen from
linux-specific changes.

This commit adds a task the only compiles the concoures binary since
that's all we need to do to satisfy this check.

Also added on_abort and on_error steps for the pr notification since we
do have timeouts on the jobs and it would be nice to notify the PR that
the checked is no longer running.

Signed-off-by: Taylor Silva <dev@taydev.net>